Added the $help command. Added the $my100stats command. This command mirrors the current $my100status command. Fixed a looping glitch with the $my100status command. Removed the $botstats command.
Enchanced the bots command checking to include a breakdown of the actual command and any associated parameters for the command.

Reorganized folder structure to include bot, helper and messages directories.
Renamed JavaScript files in the /bot/commands directory to reflect the name of the actual.

Created a "commands.process" function that will check for an issues command and break that command down into the acutal command and any associated parameters.
Created a "commands.execute" function that will check for a valid command issued to the bot.
Created a "file.read" function to read content from the local file system.
Created a "messages.format" function that can load local content from the /messages/templates directory and dynamic replace values in that file with variables passed via the function's parameters.

Added usage checking for both the $my100status, $my100stats and $stats commands.

Improved the $test command by adding a check to make sure it is being running by the test bot instead of the live bot.

Various other code optimization.